Added check for valid id on the server
Revert "Added check for valid id on the server"

This reverts commit ea934a0d4f2f3303616611505ad96db38704b902.

Revert "Revert "Added check for valid id on the server""

This reverts commit 720fb4721c45d335395ed48093b7995f749bbd4a.

Added checker for bot in premove

Added cant premove bots in premove

Fixed but

Apparently members.fetch() throws an exception if the user is not caught so I added a try/catch to handle it

the user doesn't exist on the server*

+++ src/bot/bot.ts
@@ -252,12 +252,19 @@ export default class Bot {
     });
 
     command.on('premove', async (message: Message) => {
+      if (message.author.bot) return;
       const [num, ...msg] = message.content.trim().split(' ');
       const id = parseInt(num);
 
       message.delete();
       if (isNaN(id) || !msg || !msg.length) return message.channel.send('Syntax: {prefix}premove {userId} {message}');
       if (id === parseInt(message.author.id)) return message.channel.send('You cannot premove yourself');
+      try {
+        const member = await message.guild.members.fetch(num);
+        if (member.user.bot) return message.channel.send('Can\'t premove bots');
+      } catch (e) {
+        return message.channel.send('The user you are trying to premove doesn\'t exist on this server');
+      }
 
       const content = msg.join(' ').trim();
       this.ps = await this.premoves.find();
